rm unnecessary try/catch block around the Ilios client instantiation.
the only thing that would cause an exception to get thrown in the
constructor of the Ilios client class would be emitted out of the
Moodle-core get_config() method. which would only occur if we tried to
read plugin config during system installation. we're not doing this
here, so let's get rid of it.
as a positive side-effect, this also eliminates the need to correct the incorrect
position of the chained re-thrown exception that's  being passed into
the Exception constructor.

while at it, i moved the client instantiation out of the try/catch block
in the new sync job form. should make it clearer what we're actually
tring to catch there.